requirements.txt
### google ###
gcsfs==2023.6.0
gmaps==0.9.0
google-cloud-aiplatform==1.28.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-dialogflow-cx==1.21.0
google-cloud-discoveryengine==0.11.1
google-cloud-speech==2.21.0
google-cloud-storage==2.5.0
googlemaps==4.10.0
looker_sdk==23.16.0
ortools==9.7.2996
vertexai==0.0.1

### langchain ###
aiohttp==3.8.5
db_dtypes==1.1.1
langsmith==0.0.14
mypy_extensions==1.0.0
numexpr==2.8.4
openapi-schema-pydantic==1.2.4
SQLAlchemy==1.4.49
sqlalchemy-bigquery==1.7.0

### others ###
datetime==5.1
Flask==2.2.2
functions-framework==3.3.0
ipython==7.16.1
ipywidgets==7.5.1
notebook==6.0.3
pandas==1.3.2 ## original one in demos
pydantic==1.10.12
python-dateutil==2.8.2
pytz==2020.1
pyyaml==5.3.1
seaborn==0.10.1
streamlit==1.24.0
streamlit-player==0.1.5
st-pages==0.4.3
# timedelta==2020.12.3



### ZZZ ### 
shapely==1.8.5 ## only for DocAI demo
### ZZZ ### # scann==1.2.9 ## only for DocAI demo
### ZZZ ### 
Pillow==9.5.0 ## only for DocAI demo
### ZZZ ### 
pycryptodome ## only for DocAI demo
### ZZZ ### 
pypdf==3.11.0 ## only for DocAI demo
### ZZZ ### # 
tqdm==4.66.1 ## only for DoCAI demo
### ZZZ ###
# pandas==2.0.3 ## only for DocAI demo
### ZZZ ### 
matplotlib==3.7.1  ## only for DocAI demo
### ZZZ ### 
google-cloud-documentai==2.16.0 ## only for DocAI demo


### Graveyard ### 
# looker_sdk==0.1.3b20